+++ docs/history.rst
@@ -119,6 +119,15 @@ Backwards Compatibility Notes
   have been removed, as they are redundant with the new
   ``PythonExecutable.add_python_package_distribution_resource()``, which
   accepts a ``location`` argument to control the resource location.
+* The Starlark method ``PythonExecutable.add_extension_module()`` has been
+  renamed to ``PythonExecutable.add_python_extension_module()`` and it
+  now accepts a ``location`` keyword argument defining the resource
+  location.
+* The Starlark methods ``PythonExecutable.add_in_memory_extension_module()``
+  and ``PythonExecutable.add_filesystem_relative_extension_module()`` have
+  been removed, as they are redundant with the new
+  ``PythonExecutable.add_python_extension_module()``, which accepts a
+  ``location`` argument to control the resource location.
 
 New Features
 ^^^^^^^^^^^^